+++ bulk-delete.php
@@ -5,7 +5,7 @@
  * Plugin URI: http://bulkwp.com
  * Description: Bulk delete users and posts from selected categories, tags, post types, custom taxonomies or by post status like drafts, scheduled posts, revisions etc.
  * Donate Link: http://sudarmuthu.com/if-you-wanna-thank-me
- * Version: 5.5.5
+ * Version: 5.5.6
  * License: GPL
  * Author: Sudar
  * Author URI: http://sudarmuthu.com/
@@ -14,7 +14,7 @@
  * === RELEASE NOTES ===
  * Check readme file for full release notes
  *
- * @version    5.5.5
+ * @version    5.5.6
  * @author     Sudar
  * @package    BulkDelete
  */
@@ -51,7 +51,7 @@ final class Bulk_Delete {
 	private $controller;
 
 	// version
-	const VERSION                   = '5.5.5';
+	const VERSION                   = '5.5.6';
 
 	// Numeric constants
 	const MENU_ORDER                = '26';