+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 - Enhancements
 	- Ability to delete users who have null values for a user meta field.
+	- Added custom taxonomies in System Info.
 
 ### 2017-01-28 - v5.5.5 - (Dev time: 1 hour) ###